COMP: merge warning fixes from 3.4 release branch to trunk
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=9357
git-svn-id: http://svn.slicer.org/Slicer4/trunk@9357 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Applications/CLI/OtsuThresholdSegmentation.cxx
@@ -49,7 +49,7 @@ int main( int argc, char * argv[] )
   PARSE_ARGS;
 
 // Image Types
-  typedef  short  InputPixelType;
+  typedef  unsigned long  InputPixelType;
   typedef  unsigned long  InternalPixelType;
   typedef  short  OutputPixelType;
